requirements.txt
langchain_community
tiktoken
langchainhub
langchain
flask

llama-index
langchain-astradb
python-dotenv
huggingface_hub
sentence-transformers
ctransformers
transformers

torch
#transformers==4.36.2
accelerate==0.26.1
#sentence_transformers==2.2.2
sentence_transformers
einops==0.7.0
faiss-cpu==1.7.4
#langchain==0.1.1
#langchain-community==0.0.13
unstructured==0.12.2
llama-cpp-python
llama-index-llms-huggingface
accelerate
